Update: Add POST method to the API
Define base structure of the project. Create an internal
package de-couple interactions with database from the rest of
the application. Define base class for databases, with abstract
methods to be implemented by child classes.

Define a basic class to contain an individual contact.

Add a sample config file containing info on how to run the project,
and a python runner file to simple read the config file, and pass
the required variables to the main class which will fire the flask
server.

Add new dependencies to the project. Slight modifications to the
Makefile. By default, ensure log messages are written to a file as
well as the console.